+++ scene/3d/physics/area_3d.cpp
@@ -203,7 +203,7 @@ void Area3D::_body_enter_tree(ObjectID p_id) {
 
 	E->value.in_tree = true;
 	emit_signal(SceneStringName(body_entered), node);
-	for (int i = 0; i < E->value.shapes.size(); i++) {
+	for (uint64_t i = 0; i < E->value.shapes.size(); i++) {
 		emit_signal(SceneStringName(body_shape_entered), E->value.rid, node, E->value.shapes[i].body_shape, E->value.shapes[i].area_shape);
 	}
 }
@@ -217,7 +217,7 @@ void Area3D::_body_exit_tree(ObjectID p_id) {
 	ERR_FAIL_COND(!E->value.in_tree);
 	E->value.in_tree = false;
 	emit_signal(SceneStringName(body_exited), node);
-	for (int i = 0; i < E->value.shapes.size(); i++) {
+	for (uint64_t i = 0; i < E->value.shapes.size(); i++) {
 		emit_signal(SceneStringName(body_shape_exited), E->value.rid, node, E->value.shapes[i].body_shape, E->value.shapes[i].area_shape);
 	}
 }
@@ -327,7 +327,7 @@ void Area3D::_clear_monitoring() {
 				continue;
 			}
 
-			for (int i = 0; i < E.value.shapes.size(); i++) {
+			for (uint64_t i = 0; i < E.value.shapes.size(); i++) {
 				emit_signal(SceneStringName(body_shape_exited), E.value.rid, node, E.value.shapes[i].body_shape, E.value.shapes[i].area_shape);
 			}
 
@@ -356,7 +356,7 @@ void Area3D::_clear_monitoring() {
 				continue;
 			}
 
-			for (int i = 0; i < E.value.shapes.size(); i++) {
+			for (uint64_t i = 0; i < E.value.shapes.size(); i++) {
 				emit_signal(SceneStringName(area_shape_exited), E.value.rid, node, E.value.shapes[i].area_shape, E.value.shapes[i].self_shape);
 			}
 
@@ -409,7 +409,7 @@ void Area3D::_area_enter_tree(ObjectID p_id) {
 
 	E->value.in_tree = true;
 	emit_signal(SceneStringName(area_entered), node);
-	for (int i = 0; i < E->value.shapes.size(); i++) {
+	for (uint64_t i = 0; i < E->value.shapes.size(); i++) {
 		emit_signal(SceneStringName(area_shape_entered), E->value.rid, node, E->value.shapes[i].area_shape, E->value.shapes[i].self_shape);
 	}
 }
@@ -423,7 +423,7 @@ void Area3D::_area_exit_tree(ObjectID p_id) {
 	ERR_FAIL_COND(!E->value.in_tree);
 	E->value.in_tree = false;
 	emit_signal(SceneStringName(area_exited), node);
-	for (int i = 0; i < E->value.shapes.size(); i++) {
+	for (uint64_t i = 0; i < E->value.shapes.size(); i++) {
 		emit_signal(SceneStringName(area_shape_exited), E->value.rid, node, E->value.shapes[i].area_shape, E->value.shapes[i].self_shape);
 	}
 }
